% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{build_blocklist}
\alias{build_blocklist}
\title{build list of time blocks that are correlated based on the maximum time and type of temporal model
#'}
\usage{
build_blocklist(tmax, type = "EI")
}
\description{
build list of time blocks that are correlated based on the maximum time and type of temporal model
#'
}
\keyword{internal}
