% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_phi_matrix}
\alias{build_phi_matrix}
\title{Build intermediate C(phi,n) matrix in inversion of Exchangeable variance matrix}
\usage{
build_phi_matrix(n, phi, directed = T, sta_flag = F)
}
\description{
Build intermediate C(phi,n) matrix in inversion of Exchangeable variance matrix
}
\keyword{internal}
