% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{calculate_matrix_params}
\alias{calculate_matrix_params}
\title{calculate parameter estimates for different types of matrices, i.e. 6a, 6b, etc.}
\usage{
calculate_matrix_params(ilist, e, tmax, type = "EI")
}
\description{
calculate parameter estimates for different types of matrices, i.e. 6a, 6b, etc.
}
\keyword{internal}
