% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{calculate_parameter_inverse}
\alias{calculate_parameter_inverse}
\title{Invert matrix parameters based on inputs.}
\usage{
calculate_parameter_inverse(n, tmax, params, type = "EI")
}
\description{
Invert matrix parameters based on inputs.
}
\keyword{internal}
