% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{coef.lmnet}
\alias{coef.lmnet}
\title{Coef S3 generic for class lmnet}
\usage{
\method{coef}{lmnet}(object, ...)
}
\arguments{
\item{object}{lmnet object}

\item{...}{ignored}
}
\description{
Coef S3 generic for class lmnet
}
