% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine}
\alias{combine}
\title{Find all possible combinations of elements in two vectors, or all combinations of all elements in one without repeats}
\usage{
combine(v1, v2 = NA)
}
\description{
Find all possible combinations of elements in two vectors, or all combinations of all elements in one without repeats
}
\keyword{internal}
