% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{eigen_exch}
\alias{eigen_exch}
\title{Eigenvalues of exchangeable matrices
 if calcall == TRUE, then output eigenvalues with multiplicities
 Outputs eigenvectors when directed==FALSE}
\usage{
eigen_exch(n, phi, directed = FALSE, calcall = TRUE)
}
\description{
Eigenvalues of exchangeable matrices
 if calcall == TRUE, then output eigenvalues with multiplicities
 Outputs eigenvectors when directed==FALSE
}
\keyword{internal}
