% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{eigen_exch_time}
\alias{eigen_exch_time}
\title{Compute eigenvalues of covariance matrices of jointly exchangeable errors with repeated observations}
\usage{
eigen_exch_time(n, tmax, params, directed, type)
}
\description{
Compute eigenvalues of covariance matrices of jointly exchangeable errors with repeated observations
}
\keyword{internal}
