% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GEE.est}
\alias{GEE.est}
\title{Perform GEE estimate / IRWLS of coefficients}
\usage{
GEE.est(row.list, Y, X, n, directed = TRUE, beta_start = NULL,
  missing = FALSE, dyads = NULL, tol.in = 1e-06, maxit = 10000,
  ndstop = TRUE, verbose = FALSE)
}
\description{
Perform GEE estimate / IRWLS of coefficients
}
\keyword{internal}
