% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{GEE_est_time}
\alias{GEE_est_time}
\title{Perform GEE estimate / IRWLS of coefficients for temporal data}
\usage{
GEE_est_time(Y, X, n, tmax, directed = TRUE, type = "EE",
  write_dir = NULL, missing = FALSE, tol.in = 1e-06, maxit = 10000,
  ndstop = TRUE, verbose = FALSE)
}
\description{
TO DO:
 eigenvalues / ndstop
 fix for missingness
}
\keyword{internal}
