% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_lmnet.R
\name{inputs_lmnet}
\alias{inputs_lmnet}
\title{Input preprocessing}
\usage{
inputs_lmnet(Xlist, Y = NULL, directed = TRUE, add_intercept = TRUE)
}
\arguments{
\item{Xlist}{List of \eqn{n \times n} matrices, possibly containing response matrix labeled `Y'. Diagonals (self-loops) are ignored.}

\item{Y}{Optional \eqn{n \times n} response matrix. NAs in this matrix will be automatically removed. Diagonals (self-loops) are ignored.}

\item{directed}{Optional logical indicator of whether input data is for a directed network, default is \code{TRUE}. Undirected data format is lower triangle of adjacencey matrix.}

\item{add_intercept}{Optional logical indicator of whether intercept should be added to X, default is \code{TRUE}.}
}
\value{
A list of:
\item{Y}{Vector of responses (column-wise vectorization order) of appropriate length.}
\item{X}{Matrix of covariates (column-wise vectorization order) of appropriate size.}
\item{nodes}{2-column matrix indicating directed relation pairs to which each entry in \eqn{Y} and each row in \eqn{X} corresponds.}
}
\description{
Prepare covariates and optional response in adjacency matrix form. If undirected, the values are drawn from the lower triangle of the adjacency matrices.
}
\details{
This function takes a list of network covariates (in adjacency matrix form) and prepares them for the regression code \code{lmnet}.
}
\examples{
set.seed(1)
n <- 10
Xlist <- list(matrix(rnorm(n^2),n,n), matrix(sample(c(0,1), n^2, replace=TRUE),n,n))
Xlist$Y <- matrix(rnorm(n^2), n, n)
Xlist$Y[1:5] <- NA
r <- inputs_lmnet(Xlist)
r
lmnet(r$Y,r$X,nodes=r$nodes)

}
\seealso{
\code{\link{lmnet}}, \code{\link{vhat_exch}}
}
\keyword{external}
