% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make.positive.var}
\alias{make.positive.var}
\title{Replace negative eigenvalues with zeros in variance matrix}
\usage{
make.positive.var(V.test)
}
\description{
Replace negative eigenvalues with zeros in variance matrix
}
\keyword{internal}
