% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{meatABC}
\alias{meatABC}
\title{Matrix product of A^TBC where B is a short list of parameters
 A and C must be matrices
 B is parameterized by phi, row.list, and assumed symmetric without repeats, with phi[1] along diagonal}
\usage{
meatABC(row.list, phi, A, C, directed)
}
\description{
Matrix product of A^TBC where B is a short list of parameters
 A and C must be matrices
 B is parameterized by phi, row.list, and assumed symmetric without repeats, with phi[1] along diagonal
}
\keyword{internal}
