% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{model.matrix.lmnet}
\alias{model.matrix.lmnet}
\title{model.matrix S3 generic for class lmnet}
\usage{
\method{model.matrix}{lmnet}(object, ...)
}
\arguments{
\item{object}{lmnet object}

\item{...}{ignored}
}
\description{
model.matrix S3 generic for class lmnet
}
