% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{node.set}
\alias{node.set}
\title{Generate node sets of various overlapping dyad pairs}
\usage{
node.set(n.tot, directed = T)
}
\description{
Generate node sets of various overlapping dyad pairs
}
\keyword{internal}
