% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{node_gen_time}
\alias{node_gen_time}
\title{Make complete node indices for temporal relational data}
\usage{
node_gen_time(n, directed, tmax)
}
\description{
Make complete node indices for temporal relational data
}
\keyword{internal}
