% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{node_preprocess_time}
\alias{node_preprocess_time}
\title{Pre-processes data for ordering, FOR TEMPORAL DATA, etc.}
\usage{
node_preprocess_time(Y, X, directed, nodes, tmax, type, subtract = NULL)
}
\description{
TO-DO:
   missing data
}
\keyword{internal}
