% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{param_est_single_ilist}
\alias{param_est_single_ilist}
\title{Given matrix of time blocks and a particular exchangeable parameter set (within each block),
calculate a single parameter/phi. ASSUMES NO MISSING DATA}
\usage{
param_est_single_ilist(ilist, blockmat, e, d.tot)
}
\description{
Given matrix of time blocks and a particular exchangeable parameter set (within each block),
calculate a single parameter/phi. ASSUMES NO MISSING DATA
}
\keyword{internal}
