% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{plot.lmnet}
\alias{plot.lmnet}
\title{Plot S3 generic for class lmnet}
\usage{
\method{plot}{lmnet}(x, ...)
}
\arguments{
\item{x}{lmnet object}

\item{...}{ignored}
}
\description{
Plot S3 generic for class lmnet
}
