% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{print.lmnet}
\alias{print.lmnet}
\title{Print S3 generic for class lmnet}
\usage{
\method{print}{lmnet}(x, ...)
}
\arguments{
\item{x}{lmnet object}

\item{...}{ignored}
}
\description{
Print S3 generic for class lmnet
}
