% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{print.summary.lmnet}
\alias{print.summary.lmnet}
\title{Print generic for class summary.lmnet}
\usage{
\method{print}{summary.lmnet}(x)
}
\description{
Print generic for class summary.lmnet
}
\keyword{internal}
