% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhat_exch.R
\name{print.summary.vexch}
\alias{print.summary.vexch}
\title{Print generic for summary.vexch object}
\usage{
\method{print}{summary.vexch}(x)
}
\description{
Print generic for summary.vexch object
}
\keyword{internal}
