% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnet.R
\name{print.summary.vnet}
\alias{print.summary.vnet}
\title{Print S3 generic for summary.vnet object}
\usage{
\method{print}{summary.vnet}(x, ...)
}
\arguments{
\item{x}{summary.vnet object}

\item{...}{ignored}
}
\description{
Print S3 generic for summary.vnet object
}
