% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnet.R
\name{print.vnet}
\alias{print.vnet}
\title{Print S3 generic for vnet object}
\usage{
\method{print}{vnet}(x, ...)
}
\arguments{
\item{x}{vnet object}

\item{...}{ignored}
}
\description{
Print S3 generic for vnet object
}
