% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_time.R
\name{row_list_time}
\alias{row_list_time}
\title{Make row list for complete temporal relational data}
\usage{
row_list_time(n, directed, tmax, type)
}
\description{
Make row list for complete temporal relational data
}
\keyword{internal}
