% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Sigma.ind}
\alias{Sigma.ind}
\title{Generate list indicator matrix of overlapping dyads}
\usage{
Sigma.ind(n.tot, directed, sta_flag = F)
}
\description{
Generate list indicator matrix of overlapping dyads
}
\keyword{internal}
