% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{summary.lmnet}
\alias{summary.lmnet}
\title{Summary S3 generic for class lmnet}
\usage{
\method{summary}{lmnet}(object, ...)
}
\arguments{
\item{object}{lmnet object}

\item{...}{ignored}
}
\description{
Summary S3 generic for class lmnet
}
