% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnet.R
\name{summary.vnet}
\alias{summary.vnet}
\title{Summary S3 generic for vnet object}
\usage{
\method{summary}{vnet}(object, ...)
}
\arguments{
\item{object}{vnet object}

\item{...}{ignored}
}
\description{
Summary S3 generic for vnet object
}
