% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{symm_square_root}
\alias{symm_square_root}
\title{Compute symmetric square root of A, assuming it is real, symmetric, positive definite}
\usage{
symm_square_root(A, prec = 15)
}
\description{
Compute symmetric square root of A, assuming it is real, symmetric, positive definite
}
\keyword{internal}
