% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{vcov.lmnet}
\alias{vcov.lmnet}
\title{vcov S3 generic for class lmnet}
\usage{
\method{vcov}{lmnet}(object, ...)
}
\arguments{
\item{object}{lmnet object}

\item{...}{ignored}
}
\description{
vcov S3 generic for class lmnet
}
