% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnet.R
\name{print.summary.lmnet}
\alias{print.summary.lmnet}
\title{Print S3 generic for class summary.lmnet}
\usage{
\method{print}{summary.lmnet}(x, ...)
}
\arguments{
\item{x}{summary.lmnet object}

\item{...}{ignored}
}
\description{
Print S3 generic for class summary.lmnet
}
